configured app for mysql db

+++ src/test/java/com/springframework/domain/CategoryTest.java
@@ -16,7 +16,8 @@ public class CategoryTest {
 
     @Test
     public void getId() {
-        Long idValue = 4L;
+//        Long idValue = 4L;
+        Integer idValue = 4;
 
         category.setId(idValue);
